BK-4756: Populate request body for urlencoded content type
When creating a hammock request the body is not populated when the
backend request is an 'application/x-www-form-urlencoded' content
type.

The body from the backend request is exposed as params for the
urlencoded content type.  The body content should be reconstituted as
a string as per the original backend request.

Based on the content type the reconstituted string or the original
backend request stream will be passed into the constructor of the
hammock request.

Added a constant for the urlencoded type and a unit test.